change nng_msg sending from sunneed_listener to fix invalid write on memmove stemming from msg_insert now pass 0 as len to nng_msg_alloc & use nng_msg_append rather than nng_msg_insert removed NNG_MSG_LEN_FIX call in sunneed_client.c as it would make length longer than buffer -- this might break something down the line? maybe? no errors from valgrind or ASAN w/ any number of devices registered now
