+++ .gitignore
@@ -41,6 +41,7 @@
 *.su
 *.idb
 *.pdb
+.vscode/*
 
 # Kernel Module Compile Results
 *.mod*
