+++ .gitmodules
@@ -7,3 +7,6 @@
 [submodule "test/munit"]
 	path = test/munit
 	url = https://github.com/nemequ/munit.git
+[submodule "ext/SunneeD_dev_drivers"]
+	path = ext/SunneeD_dev_drivers
+	url = https://github.com/jonterry9/SunneeD_dev_drivers.git
